Added some logging for the database, changed select sql statement.
Changed the select statement to match the first link stored in the
database regardless of the page number.

In read_db this is just going to match ....45-1-0-1-risitas.... or
....2-1-0-1-risitas....

In update_db the page_link stored in the database will be matched like
mentionned above, and the page link inserted into the database will be
whatever was provided in the command-line which is not important
considering what we mentionned above.