Renamed messages to posts and added self.post_added back
This flag is necessary to set the post cursor to 0 if the n + 1 page
doesn't contains a risitas post otherwise, the post cursor would refer
to a cursor that was valif but for the n - m pages where n is the
total of the pages and m is the last page where a risitas post was
found.

Probably a way to use self.past_message though.